// src/Types/RelatorioDeVendas.ts
export type RelatorioCC = {
  Loja_Compra: string | null;
  Tipo_de_Loja: string;
  Vendedor_Compra: string | null;
  Data_Compra: string;
  Nome_Cliente: string | null;
  CPF_Cliente: string | null;
  Telefone_Contato: string | null;
  Vendedor_Contato: string | null;
  Vendedor_Contato_Matricula: string | null;
  Loja_Vendedor_Contato: string | null;
  Data_Contato: string;
  Nome_Acao: string | null;
  Nome_Campanha: string | null;
  ComprasTotais: number;
};